various change settings

+++ Assets/Scripts/Enemy/EnemyStatus.cs
@@ -2,8 +2,8 @@ using Scrips.Models;
 public class EnemyStatus
 {
     public string EnemyName { get; set; }
-    public int Hp { get; set; }
-    public int Speed { get; set; }
-    public int Attack { get; set; }
+    public float Hp { get; set; }
+    public float Speed { get; set; }
+    public float Attack { get; set; }
     public Constans.EnemyType EnemyType { get; set; }
 }
